refactor(linux-loader): split init_stack functionality into separate module
- Move init_stack method and related structures to new stack.rs module
- Update imports and remove redundant code from loader.rs
- This change prepares for integrating Linux-specific functionality
